services/inference/requirements.txt
google-cloud-monitoring
fastapi
uvicorn[standard]
pydantic
